# src/main/resources/static/apidoc/swagger.json
{"swagger":"2.0","info":{"description":"Spring Demo","version":"1.0","title":"Spring Demo","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"horse-race-controller","description":"Horse Race Controller"},{"name":"pdf-generator-controller","description":"PDF Generator Controller"},{"name":"spring-micro-controller","description":"Spring Micro Controller"}],"paths":{"/api/createPdf":{"post":{"tags":["pdf-generator-controller"],"summary":"createPdf","operationId":"createPdfUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"param","description":"param","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/getExistingPdf":{"get":{"tags":["pdf-generator-controller"],"summary":"getExistingPdf","operationId":"getExistingPdfUsingGET","produces":["*/*"],"parameters":[{"name":"filename","in":"query","description":"filename","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Resource"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/getFile":{"get":{"tags":["spring-micro-controller"],"summary":"download","operationId":"downloadUsingGET","produces":["*/*"],"parameters":[{"name":"param","in":"query","description":"param","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Resource"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/getPdf":{"post":{"tags":["pdf-generator-controller"],"summary":"getPdf","operationId":"getPdfUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"param","description":"param","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Resource"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/hello":{"get":{"tags":["spring-micro-controller"],"summary":"getMessage","operationId":"getMessageUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/horseReset":{"get":{"tags":["horse-race-controller"],"summary":"getHorseReset","operationId":"getHorseResetUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/horseprices":{"get":{"tags":["horse-race-controller"],"summary":"getHorsePriceStream","operationId":"getHorsePriceStreamUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Horse"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/horserace":{"get":{"tags":["horse-race-controller"],"summary":"getHorseRaceStream","operationId":"getHorseRaceStreamUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Horse"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/stock":{"get":{"tags":["spring-micro-controller"],"summary":"getDataStream","operationId":"getDataStreamUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Stock"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/stream-flux":{"get":{"tags":["spring-micro-controller"],"summary":"streamEvents","operationId":"streamEventsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Flux«ServerSentEvent«string»»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"File":{"type":"object","properties":{"absolute":{"type":"boolean"},"absoluteFile":{"$ref":"#/definitions/File"},"absolutePath":{"type":"string"},"canonicalFile":{"$ref":"#/definitions/File"},"canonicalPath":{"type":"string"},"directory":{"type":"boolean"},"file":{"type":"boolean"},"freeSpace":{"type":"integer","format":"int64"},"hidden":{"type":"boolean"},"name":{"type":"string"},"parent":{"type":"string"},"parentFile":{"$ref":"#/definitions/File"},"path":{"type":"string"},"totalSpace":{"type":"integer","format":"int64"},"usableSpace":{"type":"integer","format":"int64"}},"title":"File"},"Flux«ServerSentEvent«string»»":{"type":"object","properties":{"prefetch":{"type":"integer","format":"int32"}},"title":"Flux«ServerSentEvent«string»»"},"Horse":{"type":"object","properties":{"distanceTravelled":{"type":"integer","format":"int64"},"hasFinished":{"type":"boolean"},"jockey":{"type":"string"},"name":{"type":"string"},"number":{"type":"string"},"position":{"type":"integer","format":"int64"},"price":{"type":"string"}},"title":"Horse"},"InputStream":{"type":"object","title":"InputStream"},"Resource":{"type":"object","properties":{"description":{"type":"string"},"file":{"$ref":"#/definitions/File"},"filename":{"type":"string"},"inputStream":{"$ref":"#/definitions/InputStream"},"open":{"type":"boolean"},"readable":{"type":"boolean"},"uri":{"$ref":"#/definitions/URI"},"url":{"$ref":"#/definitions/URL"}},"title":"Resource"},"ServerSentEvent«string»":{"type":"object","title":"ServerSentEvent«string»"},"Stock":{"type":"object","properties":{"name":{"type":"string"},"price":{"type":"string"}},"title":"Stock"},"URI":{"type":"object","properties":{"absolute":{"type":"boolean"},"authority":{"type":"string"},"fragment":{"type":"string"},"host":{"type":"string"},"opaque":{"type":"boolean"},"path":{"type":"string"},"port":{"type":"integer","format":"int32"},"query":{"type":"string"},"rawAuthority":{"type":"string"},"rawFragment":{"type":"string"},"rawPath":{"type":"string"},"rawQuery":{"type":"string"},"rawSchemeSpecificPart":{"type":"string"},"rawUserInfo":{"type":"string"},"scheme":{"type":"string"},"schemeSpecificPart":{"type":"string"},"userInfo":{"type":"string"}},"title":"URI"},"URL":{"type":"object","properties":{"authority":{"type":"string"},"content":{"type":"object"},"defaultPort":{"type":"integer","format":"int32"},"file":{"type":"string"},"host":{"type":"string"},"path":{"type":"string"},"port":{"type":"integer","format":"int32"},"protocol":{"type":"string"},"query":{"type":"string"},"ref":{"type":"string"},"userInfo":{"type":"string"}},"title":"URL"}}}